requirements.txt
openai-whisper
tensorflow
pydub
librosa
numpy
pandas
nltk
transformers